% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CamBioR.r
\docType{package}
\name{CamBioR}
\alias{CamBioR}
\title{\pkg{CamBioR}}
\description{
\pkg{CamBioR} is an R package for holding all the reuasable bioinformatics R code I write.
}
\details{
\pkg{CamBioR} is a package for
}
\section{NA}{

\itemize{
\item \code{\link{small_unittest.fasta}} An example file in fasta format.
\item \code{\link{small_unittest.fastq}} An example file in fastq format.
}
}

\section{Data}{

\itemize{
\item \code{\link{small_unittest.fasta}} An example file in fasta format.
\item \code{\link{small_unittest.fastq}} An example file in fastq format.
}
}

\author{
Cameron M. Nugent
}
