% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_fasta}
\alias{read_fasta}
\title{Read in raw data from a fasta file.}
\usage{
read_fasta(filename)
}
\arguments{
\item{filename}{The name of the fasta file to read data from.}
}
\value{
returns a dataframe with the columns: 
"header", "sequence"
}
\description{
Read in raw data from a fasta file.
}
\examples{
fastq_example_file = system.file('extdata/small_unittest.fasta', 
                                  package = 'bioreadr')
data = read_fastq(fastq_example_file)
}
