% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_fastq}
\alias{read_fastq}
\title{Read in data from a fastq file.}
\usage{
read_fastq(filename)
}
\arguments{
\item{filename}{The name of the fastq file to read data from.}
}
\value{
returns a dataframe with the columns: 
"header", "sequence", "strand", "qual"
}
\description{
Read in data from a fastq file.
}
\examples{
fastq_example_file = system.file('extdata/small_unittest.fastq', 
                                  package = 'bioreadr')
data = read_fastq(fastq_example_file)
}
