% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_fasta}
\alias{write_fasta}
\title{Write data to a fasta file.}
\usage{
write_fasta(df, path)
}
\arguments{
\item{df}{The dataframe of sequence data to write to file. The dataframe must
contain the columns: "header" and "sequence". Other columns in the dataframe
are permitted but will be ignored.}

\item{path}{Path or connection to write to.}
}
\description{
Write data to a fasta file.
}
\examples{
fasta_ex_dat = data.frame(
    header = c("seq1", "seq2", "seq3"),
    sequence = c("ATGC", "AATTGGCC", "TGACTGAC"),
    stringsAsFactors = FALSE
    )
data = write_fasta(fasta_ex_dat, "example_output.fasta")
}
