% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_fastq}
\alias{write_fastq}
\title{Write data to a fastq file.}
\usage{
write_fastq(df, path)
}
\arguments{
\item{df}{The dataframe of sequence data to write to file. The dataframe must
contain the columns: "header", "sequence", "strand", and "qual". 
Other columns in the dataframe are permitted but will be ignored.}

\item{path}{Path or connection to write to.}
}
\description{
Write data to a fastq file.
}
\examples{
fastq_ex_dat = data.frame(
    header = c("seq1", "seq2", "seq3"),
    sequence = c("ATGC", "AATTGGCC", "TGACTGAC"),
    strand = c("+", "+","+"),
    qual = c("~~ua", "~AZ~ngZO", "lvykl~Qv"),
    stringsAsFactors = FALSE)
data = write_fastq(fastq_ex_dat, "example_output.fastq")
}
